Stop mis-interpreting the 'removable' node in xenstore...
... and the VDISK_REMOVABLE bit in the value of the 'info' node.

They both apply to the media and not the device itself. PV devices are always
removable.
The comment in libxl_disk.c concerning the 'removable' flag states that:

"Currently there is only one removable device -- CDROM"

This is not conclusive but it is reasonable to infer from that the removabilty
refers to the media and not the drive itself. (CDROM drives in typical servers
are not removable).
The code in Linux xen-blkback/xenbus.c sets VDISK_REMOVABLE if the underlying
block device has the GENHD_FL_REMOVABLE flag, and the comment above the
definition of that flag in genhd.h states:

"``GENHD_FL_REMOVABLE`` (0x0001): indicates that the block device gives access
to removable media.
When set, the device remains present even when media is not inserted.
Must not be set for devices which are removed entirely when the media is
removed."

This patch, therefore, stops using these values to indicate the removability
of the target devices and instead uses them as they were intended, to indicate
the removability of the media.

NOTE: The code in XENCRSH is modified to simply ignore the 'removable' node.
      The value currently sampled is not used.
      The Removable BOOLEAN field currently in XENVBD_CAPS is also moved into
      XENVBD_FEATURES for consistency with other values sampled from xenstore
      nodes.

Signed-off-by: Paul Durrant <pdurrant@amazon.com>