Clear features exposed by backend on disconnect
By clearing the backend features, it will be possible to migrate to
a backend that exposes different features. This will allow migration
to/from backends that expose different enablement for Barrier, Flush,
and Discard operations without relying on the backend to return
BLKIF_OP_ENOTSUPP to disable an unsupported feature.

Also removes the check for change while reading disk features, as this
is only called once during connection, and always log the values read.

Signed-off-by: Owen Smith <owen.smith@cloud.com>