Clear operations exposed by backend on disconnect
By clearing the backend operations, it will be possible to live migrate
to a backend that exposes different operations. This will allow live migration
to/from backends that expose different enablement for Barrier, Flush,
and Discard operations without relying on the backend to return
BLKIF_RSP_EOPNOTSUPP to disable an unsupported operation.

Also removes the check for change while reading disk features and always
log the values read, as this is only called once during connection.

Signed-off-by: Owen Smith <owen.smith@cloud.com>